Move frontend.c into retroarch.c
