+++ griffin/griffin.c
@@ -1066,11 +1066,6 @@ UI
 #endif
 
 /*============================================================
-MAIN
-============================================================ */
-#include "../frontend/frontend.c"
-
-/*============================================================
 GIT
 ============================================================ */
